Custom sorting
Work notes:

Display the current folder path (breadcrumbs) as seen in the doc manager to give user context. Unlike the doc manager these breadcrumbs are not clickable. The breadcrumbs are placed in the top bar beside the title "Default Sort Order".

The second section lays out the sorting controls (date/name; asc/dec; apply to children; custom sort. Layout is mix of horizontal and vertical to help user understand what they can do while conserving space for the display of folders and documents in sorted order.

Next section contains two components: list of folders and list of documents. Unlike the doc manager this UI will show a smaller selection of folders and documents at the same time. This is required to (a) let the user see the effect on both folders and documents when they choose a sort option and (b) provide a means to drag and drop either folders or documents without having to scroll the UI.

Inside each of these drag and drop areas we have a sub-title "Folders" or "Documents" and the scrollable lists. There is no header section (Name, Date) because (a) the sorting controls are in the section above and (b) they have to apply to both sections; folders and documents.

The drag and drop collections are disabled if the sorting option is anything but custom.

Selecting items is only enabled if custom is selected.

If custom is selected and some items are selected these items are de-selected when the user changes the sorting option away from custom.

If the user changes the sorting option away from custom they are first given a confirmation dialog. If they select cancel they keep the custom sort order. If they select to proceed with the change the folder and document lists are immediately sorted based on the selected sort option (this is nearly instant as the sorting is done entirely on the client side).

If the sort option was any of the Name or Date choices the "Apply to children" checkbox label reflects the sorting choice. For example "Apply name ascending to all subfolders". This clearly informs the user as to what action this checkbox will invoke when they Save.

If the user selects custom and the "Apply to subfolders" option was checked then the code un-checks the "Apply to subfolders" option.
While the Custom option is selected the "Apply to subfolders" option is disabled and it's label is disabled.

When Custom is active and the user starts a drag operation on one of the lists, say folders, then the other list's drag and drop is disabled. (Otherwise the user could drop a folder into documents or visa versa.)

Initialization of the UI is based on the saved sorting options for the folder. Includes set the sorting option controls based on the incoming setting and includes disabling the Apply to Subfolders if custom is selected or update the label of the Apply control based on the Name/Date selection.  Of course, sort the lists by the sorting option.

On Save, if the Apply to Subfolders option is checked then confirm the action with the user. If they cancel return to the UI. Otherwise:
* save the sorting options for the folder
* apply to all children if checked.
* if custom sorting then update the sort order of all folders and documents

Also note that as new documents and folders are added they always will be at the end of custom sort lists. 

AT THIS TIME, the apply to all sub folders does not actually make the change to subfolders.